BETA: More event listeners: 	* Re-implemented events: 		* onDlgAnswer (Fired whenever player answer on a dialog request) 		* onPlayerLogin (Fired whenever player logs in) 		* onPlayerLogout (Fired whenever player is removed from the world) 		* onLevelChange (Fired whenever player's level has been changed) 	* Implemented new annotations giving ability to filter a bit event listeners (For now they can't be used with L2Script) 		* Message (Can be used with onDlgAnswer to filter by message id) 		* PlayerOnly (Can be used with most of event listeners to listen only for player actions) 		* SkillId (Can be used with onSkillUse to filter by skill id) 		* SkillLevel (Can be used with onSkillUse to filter by skill level) 	* Cleaned up Base64 class from useless try/catch using StandardCharsets.UTF8 	* Cleaned up EnterWorld packet. 	* Fixed wrong default type of getLong method in StatsSet it must be long not int. 	* Reworking a little old listeners no more public access of player variable. 	* Reviewed by: Zoey76, BiggBoss, Nos

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/StatsSet.java
@@ -411,7 +411,7 @@ public class StatsSet
 	 * @param deflt : long designating the default value if value associated with the key is null
 	 * @return long : value associated to the key
 	 */
-	public long getLong(String name, int deflt)
+	public long getLong(String name, long deflt)
 	{
 		Object val = _set.get(name);
 		if (val == null)
